services/recommendation-service/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
psycopg2-binary>=2.9,<3.0
PyJWT>=2.9,<3.0
requests>=2.32,<3.0
torch>=2.2,<2.7
numpy>=1.26,<3.0
matplotlib>=3.8,<4.0
scikit-learn>=1.4,<2.0
